Move HarvesterPreviewEnvironment to preview.ts
