+++ .werft/platform-delete-preview-environments-cron.ts
@@ -2,10 +2,8 @@ import { Werft } from "./util/werft";
 import * as Tracing from "./observability/tracing";
 import { SpanStatusCode } from "@opentelemetry/api";
 import { exec } from "./util/shell";
-import { previewNameFromBranchName } from "./util/preview";
-import { CORE_DEV_KUBECONFIG_PATH, HARVESTER_KUBECONFIG_PATH, PREVIEW_K3S_KUBECONFIG_PATH } from "./jobs/build/const";
-import { deleteDNSRecord } from "./util/gcloud";
-import * as VM from "./vm/vm";
+import { CORE_DEV_KUBECONFIG_PATH, HARVESTER_KUBECONFIG_PATH } from "./jobs/build/const";
+import { HarvesterPreviewEnvironment, PreviewEnvironment } from "./util/preview";
 
 // for testing purposes
 // if set to 'true' it shows only previews that would be deleted
@@ -47,189 +45,6 @@ Tracing.initialize()
         werft.endAllSpans();
     });
 
-export class HarvesterPreviewEnvironment {
-    // The prefix we use for the namespace
-    static readonly namespacePrefix: string = "preview-";
-
-    // The name of the namespace that the VM and related resources are in, e.g. preview-my-branch
-    namespace: string;
-
-    // The name of the preview environment, e.g. my-branch
-    name: string;
-
-    // The namespace in the k3s cluster where all resources are (default)
-    k3sNamespace: string = "default";
-
-    constructor(namespace: string) {
-        this.namespace = namespace;
-        this.name = namespace.replace(HarvesterPreviewEnvironment.namespacePrefix, "");
-    }
-
-    async delete(): Promise<void> {
-        VM.deleteVM({ name: this.name });
-    }
-
-    async removeDNSRecords(sliceID: string) {
-        werft.log(sliceID, "Deleting harvester related DNS records for the preview environment");
-        await Promise.all([
-            deleteDNSRecord(
-                "A",
-                `*.ssh.ws.${this.name}.preview.gitpod-dev.com`,
-                "gitpod-core-dev",
-                "preview-gitpod-dev-com",
-                sliceID,
-            ),
-            deleteDNSRecord(
-                "A",
-                `*.ws.${this.name}.preview.gitpod-dev.com`,
-                "gitpod-core-dev",
-                "preview-gitpod-dev-com",
-                sliceID,
-            ),
-            deleteDNSRecord(
-                "A",
-                `*.${this.name}.preview.gitpod-dev.com`,
-                "gitpod-core-dev",
-                "preview-gitpod-dev-com",
-                sliceID,
-            ),
-            deleteDNSRecord(
-                "A",
-                `${this.name}.preview.gitpod-dev.com`,
-                "gitpod-core-dev",
-                "preview-gitpod-dev-com",
-                sliceID,
-            ),
-            deleteDNSRecord(
-                "A",
-                `prometheus-${this.name}.preview.gitpod-dev.com`,
-                "gitpod-core-dev",
-                "preview-gitpod-dev-com",
-                sliceID,
-            ),
-            deleteDNSRecord(
-                "TXT",
-                `prometheus-${this.name}.preview.gitpod-dev.com`,
-                "gitpod-core-dev",
-                "preview-gitpod-dev-com",
-                sliceID,
-            ),
-            deleteDNSRecord(
-                "A",
-                `grafana-${this.name}.preview.gitpod-dev.com`,
-                "gitpod-core-dev",
-                "preview-gitpod-dev-com",
-                sliceID,
-            ),
-            deleteDNSRecord(
-                "TXT",
-                `grafana-${this.name}.preview.gitpod-dev.com`,
-                "gitpod-core-dev",
-                "preview-gitpod-dev-com",
-                sliceID,
-            ),
-        ]);
-    }
-
-    /**
-     * Checks whether a preview environment is active based on the db activity.
-     *
-     * It errs on the side of caution, so in case of connection issues etc. it will consider the
-     * preview environment active.
-     */
-    isActive(): boolean {
-        const sliceID = SLICES.CHECKING_FOR_DB_ACTIVITY;
-        try {
-            try {
-                VM.get({ name: this.name });
-            } catch (e) {
-                if (e instanceof VM.NotFoundError) {
-                    werft.log(
-                        sliceID,
-                        `${this.name} - is-active=false - The VM doesn't exist, deleting the environment`,
-                    );
-                    return false;
-                }
-                werft.log(
-                    sliceID,
-                    `${this.name} - is-active=true - Unexpected error trying to get the VM. Marking env as active: ${e.message}`,
-                );
-                return true;
-            }
-
-            // The preview env is its own k3s cluster, so we need to get the kubeconfig for it
-            VM.startSSHProxy({ name: this.name, slice: sliceID });
-            exec("sleep 5", { silent: true, slice: sliceID });
-
-            VM.copyk3sKubeconfig({ name: this.name, timeoutMS: 1000 * 60 * 3, slice: sliceID });
-            const kubectclCmd = `KUBECONFIG=${PREVIEW_K3S_KUBECONFIG_PATH} kubectl --insecure-skip-tls-verify`;
-
-            werft.log(sliceID, `${this.name} (${this.k3sNamespace}) - Checking status of the MySQL pod`);
-            const statusDB = exec(
-                `${kubectclCmd} get pods mysql-0 -n ${this.k3sNamespace} -o jsonpath='{.status.phase}'`,
-                { slice: sliceID, dontCheckRc: true },
-            );
-            const statusDbContainer = exec(
-                `${kubectclCmd} get pods mysql-0 -n ${this.k3sNamespace} -o jsonpath='{.status.containerStatuses.*.ready}'`,
-                { slice: sliceID, dontCheckRc: true },
-            );
-
-            if (statusDB.code != 0 || statusDB != "Running" || statusDbContainer == "false") {
-                werft.log(
-                    sliceID,
-                    `${this.name} (${this.k3sNamespace}) - is-active=false - The database is not reachable, assuming env is not active`,
-                );
-                VM.stopKubectlPortForwards();
-                exec(`rm ${PREVIEW_K3S_KUBECONFIG_PATH}`, { silent: true, slice: sliceID });
-                return false;
-            }
-
-            const dbPassword = exec(
-                `${kubectclCmd} get secret db-password -n ${this.k3sNamespace} -o jsonpath='{.data.mysql-root-password}' | base64 -d`,
-                { silent: true },
-            ).stdout.trim();
-
-            // MySQL runs in the preview env cluster that is not reachable form the job's pod, so we have to port forward
-            exec(`${kubectclCmd} -n ${this.k3sNamespace} port-forward svc/mysql 33061:3306`, {
-                async: true,
-                silent: true,
-                slice: sliceID,
-                dontCheckRc: true,
-            });
-            exec("sleep 5", { silent: true, slice: sliceID });
-
-            // Using MYSQL_PWD instead of a flag for the pwd suppresses "[Warning] Using a password on the command line interface can be insecure."
-            const dbConn = `MYSQL_PWD=${dbPassword} mysql --host=127.0.0.1 --port=33061 --user=root --database=gitpod -s -N`;
-            const active = isDbActive(this, dbConn, sliceID);
-
-            // clean after ourselves, as we'll be running this for quite a few environments
-            VM.stopKubectlPortForwards();
-            exec(`rm ${PREVIEW_K3S_KUBECONFIG_PATH}`, { silent: true, slice: sliceID });
-
-            return active;
-        } catch (err) {
-            // cleanup in case of an error
-            VM.stopKubectlPortForwards();
-            exec(`rm ${PREVIEW_K3S_KUBECONFIG_PATH}`, { silent: true, slice: sliceID });
-            werft.log(
-                sliceID,
-                `${this.name} (${this.k3sNamespace}) - is-active=true - Unable to check DB activity, assuming env is active`,
-            );
-            return true;
-        }
-    }
-
-    /**
-     * Given a branch name it will return the expected namespace of the preview environment
-     */
-    static expectedNamespaceFromBranch(branch: string): string {
-        const previewName = previewNameFromBranchName(branch);
-        return `${HarvesterPreviewEnvironment.namespacePrefix}${previewName}`;
-    }
-}
-
-export type PreviewEnvironment = HarvesterPreviewEnvironment;
-
 async function getAllPreviewEnvironments(slice: string): Promise<PreviewEnvironment[]> {
     const harvesterPreviewEnvironments = exec(
         `kubectl --kubeconfig ${HARVESTER_KUBECONFIG_PATH} get ns -o=custom-columns=:metadata.name | grep preview-`,
@@ -237,7 +52,7 @@ async function getAllPreviewEnvironments(slice: string): Promise<PreviewEnvironm
     )
         .stdout.trim()
         .split("\n")
-        .map((namespace) => new HarvesterPreviewEnvironment(namespace.trim()));
+        .map((namespace) => new HarvesterPreviewEnvironment(werft, namespace.trim()));
 
     werft.currentPhaseSpan.setAttributes({
         "preview_environments.counts.harvester": harvesterPreviewEnvironments.length,
@@ -374,7 +189,9 @@ async function determineStalePreviewEnvironments(options: {
 
     werft.log(SLICES.CHECKING_FOR_DB_ACTIVITY, `Checking ${previews.length} preview environments for DB activity`);
     const previewNamespacesWithNoDBActivity = new Set(
-        previews.filter((preview) => !preview.isActive()).map((preview) => preview.namespace),
+        previews
+            .filter((preview) => !preview.isActive(SLICES.CHECKING_FOR_DB_ACTIVITY))
+            .map((preview) => preview.namespace),
     );
 
     werft.done(SLICES.CHECKING_FOR_DB_ACTIVITY);
@@ -423,46 +240,50 @@ async function removePreviewEnvironment(previewEnvironment: PreviewEnvironment) 
 }
 
 async function removeOrphanCertificates() {
-    const certificatesNamespace = "certs"
+    const certificatesNamespace = "certs";
     werft.phase(SLICES.DELETING_ORPHAN_CERTIFICATES);
 
     const certificates = exec(
         `kubectl --kubeconfig ${CORE_DEV_KUBECONFIG_PATH} get certificates -n ${certificatesNamespace} -o=custom-columns=:metadata.name | grep harvester-`,
         { slice: SLICES.DELETING_ORPHAN_CERTIFICATES, silent: true, async: false },
     )
-    .stdout.trim()
-    .split("\n");
+        .stdout.trim()
+        .split("\n");
 
     const previews = exec(
         `kubectl --kubeconfig ${HARVESTER_KUBECONFIG_PATH} get ns -o=custom-columns=:metadata.name | grep preview-`,
         { slice: SLICES.DELETING_ORPHAN_CERTIFICATES, silent: true, async: false },
     )
-    .stdout.trim()
-    .replace(/preview-/g, "")
-    .split("\n");
+        .stdout.trim()
+        .replace(/preview-/g, "")
+        .split("\n");
 
-    certificates.forEach(certificate => {
+    certificates.forEach((certificate) => {
         const owner = exec(
             `kubectl --kubeconfig ${CORE_DEV_KUBECONFIG_PATH} get certificates ${certificate} -n ${certificatesNamespace} -o=custom-columns=:metadata.annotations.preview/owner`,
             { slice: SLICES.DELETING_ORPHAN_CERTIFICATES, silent: true, async: false },
-            ).stdout.trim()
+        ).stdout.trim();
 
         if (DRY_RUN) {
             if (!previews.includes(owner)) {
-                werft.log(SLICES.DELETING_ORPHAN_CERTIFICATES, `Certificate ${certificate} would have been deleted`)
+                werft.log(SLICES.DELETING_ORPHAN_CERTIFICATES, `Certificate ${certificate} would have been deleted`);
             } else {
-                werft.log(SLICES.DELETING_ORPHAN_CERTIFICATES, `Certificate ${certificate} would NOT have been deleted`)
+                werft.log(
+                    SLICES.DELETING_ORPHAN_CERTIFICATES,
+                    `Certificate ${certificate} would NOT have been deleted`,
+                );
             }
         }
 
         if (!previews.includes(owner) && !DRY_RUN) {
-            exec(`kubectl --kubeconfig ${CORE_DEV_KUBECONFIG_PATH} -n ${certificatesNamespace} delete --ignore-not-found=true cert ${certificate}`,
+            exec(
+                `kubectl --kubeconfig ${CORE_DEV_KUBECONFIG_PATH} -n ${certificatesNamespace} delete --ignore-not-found=true cert ${certificate}`,
                 { slice: SLICES.DELETING_ORPHAN_CERTIFICATES, async: true, silent: true },
             );
         }
     });
 
-    werft.done(SLICES.DELETING_ORPHAN_CERTIFICATES)
+    werft.done(SLICES.DELETING_ORPHAN_CERTIFICATES);
 }
 
 async function cleanLoadbalancer() {
@@ -511,43 +332,4 @@ function getAllBranches(): string[] {
     )
         .stdout.trim()
         .split("\n");
-}
-
-/**
- * Determines if the db of a preview environment is active
- * by looking if there were relevant entries in the workspace and user tables in the last 48h
- *
- */
-function isDbActive(previewEnvironment: PreviewEnvironment, dbConn: string, sliceID: string): boolean {
-    const timeout = 48;
-    let isActive = false;
-
-    const queries = {
-        d_b_workspace_instance: `SELECT TIMESTAMPDIFF(HOUR, creationTime, NOW()) FROM d_b_workspace_instance WHERE creationTime > DATE_SUB(NOW(), INTERVAL '${timeout}' HOUR) ORDER BY creationTime DESC LIMIT 1`,
-        "d_b_user-created": `SELECT TIMESTAMPDIFF(HOUR, creationDate, NOW()) FROM d_b_user WHERE creationDate > DATE_SUB(NOW(), INTERVAL '${timeout}' HOUR) ORDER BY creationDate DESC LIMIT 1`,
-        "d_b_user-modified": `SELECT TIMESTAMPDIFF(HOUR, _lastModified, NOW()) FROM d_b_user WHERE _lastModified > DATE_SUB(NOW(), INTERVAL '${timeout}' HOUR) ORDER BY _lastModified DESC LIMIT 1`,
-        d_b_workspace_instance_user: `SELECT TIMESTAMPDIFF(HOUR, lastSeen, NOW()) FROM d_b_workspace_instance_user WHERE lastSeen > DATE_SUB(NOW(), INTERVAL '${timeout}' HOUR) ORDER BY lastSeen DESC LIMIT 1`,
-    };
-
-    const result = {};
-    // let logLine = `Last Activity (hours ago):`
-    for (const [key, query] of Object.entries(queries)) {
-        // explicitly set to null, so we get an output in the logs for those queries
-        result[key] = null;
-        const queryResult = exec(`${dbConn} --execute="${query}"`, { silent: true, slice: sliceID });
-        if (queryResult.length > 0) {
-            result[key] = queryResult.stdout.trim();
-            isActive = true;
-        }
-    }
-
-    const logLines = Object.entries(result).map((kv) => `${kv.join(":")}`);
-    const logLine = `Last Activity (hours ago): ${logLines.join(",")}`;
-
-    werft.log(
-        sliceID,
-        `${previewEnvironment.name} (${previewEnvironment.namespace}) - is-active=${isActive} ${logLine}`,
-    );
-
-    return isActive;
 }
